Transation base created
The transaction has been created, but the crypto input has been changed, not the output.

So, if I buy some Bitcoin using Ethereum, the Bitcoin will be added to my wallet, but the Ethereum will not subtract from my wallet.

furthermore, I do not make the conversion off the cryptos, and I do not work with moneys, I just cryptos